requirements.txt
# ===========================================
# MetaLayer Production Requirements
# ===========================================
# Astro Runtime includes many pre-installed providers: 
# https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages

# ===========================================
# DATABASE PROVIDERS (PostgreSQL Only)
# ===========================================
apache-airflow-providers-postgres==5.11.1
apache-airflow-providers-common-sql==1.18.1
psycopg2-binary==2.9.10

# ===========================================
# DATA PROCESSING LIBRARIES
# ===========================================
pandas==2.2.3
SQLAlchemy==2.0.36
numpy==1.26.4

# ===========================================
# DATA QUALITY & MONITORING
# ===========================================
great-expectations==1.0.0
ruamel.yaml==0.18.6

# ===========================================
# PERFORMANCE OPTIMIZATION
# ===========================================
psycopg2-binary==2.9.10
sqlparse==0.5.1

# ===========================================
# DATA QUALITY & VALIDATION
# ===========================================
great-expectations==0.18.19
pydantic==2.5.3

# ===========================================
# MONITORING & OBSERVABILITY
# ===========================================
prometheus-client==0.19.0

# ===========================================
# UTILITIES
# ===========================================
python-dateutil==2.8.2
pytz==2024.1
